// src/lib/api/models/updateAppointmentDtoStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type UpdateAppointmentDtoStatus = typeof UpdateAppointmentDtoStatus[keyof typeof UpdateAppointmentDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateAppointmentDtoStatus = {
  SCHEDULED: 'SCHEDULED',
  CONFIRMED: 'CONFIRMED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  NO_SHOW: 'NO_SHOW',
} as const;